+++ source/include/multiplexer.h
@@ -6,12 +6,12 @@
 
 namespace Multiplexer {
     void begin(
-        int32_t s0Pin,
-        int32_t s1Pin,
-        int32_t s2Pin,
-        int32_t s3Pin
+        uint8_t s0Pin,
+        uint8_t s1Pin,
+        uint8_t s2Pin,
+        uint8_t s3Pin
     );
     // No need to stop anything here since once it executes at the beginning, there is no other use for this
     
-    void setChannel(uint32_t channel);
+    void setChannel(uint8_t channel);
 }
